Modified on 17-08-2024

+++ src/ZZZZZ/Accenture/CanopyArea.java
@@ -5,9 +5,9 @@ public class CanopyArea {
         int n = 7;
         System.out.println(canopyAreaCalculation(n));
     }
-    static  int canopyAreaCalculation(int n){
+    static  float canopyAreaCalculation(int n){
 
-        int area =  (3.14 * n * n);
-         return  Math.abs(area);
+        int area = (int) (3.14 * n * n);
+         return (float) Math.abs(area);
     }
 }
